Isolate more tests
